Bump 9.5 extension version to 1.0.6.
The 9.5 version of the extension was accidentally left at 1.0 which does not reflect the various updates. The default is now set to 1.0.6 and an empty update script is included. Run:

alter extension pgaudit update to '1.0.6'

to update the extension version to 1.0.6, though there are no apparent problems with leaving it at 1.0.