Differentiate UPDATE from SELECT FOR UPDATE queries
When auditing write operations (pgaudit.log = 'write'), we check
whether the query requires ACL_UPDATE permissions.  This also catches
SELECT FOR UPDATE, SELECT FOR KEY SHARE, etc. queries, because there
is no separate permission for those.  This is especially annoying
because internally generated foreign-key queries get logged.

To avoid this, in PostgreSQL 13 we can also check the rellockmode
field of the range table entry.  If it's at least RowExclusiveLock,
then it's a real UPDATE, else it's some kind of SELECT with locking
clause.