Fix logic to properly classify SELECT FOR UPDATE as SELECT.
This logic was submitted in PR #88 but there was some confusion on my part about what it was supposed to do. Since the title was "Suppress logging for internally generated foreign-key queries" I tried to make it do that, and broke SELECT FOR UPDATE logging, which unfortunately had no test.

Reading the PR again, it seems Peter's intention was only to correctly classify SELECT FOR UPDATE as SELECT. In any case that represents an improvement over what we have, even if it does not suppress logging for internally generated foreign-key queries, at least not in the case of SELECT.

So, revert the logic that suppressed the SELECT FOR UPDATE logging and use Peter's logic that correctly classifies them as SELECT.

Also add a SELECT FOR UPDATE test to prevent regressions.

Reported by Sergey Shinderuk (@shinderuk).
Reviewed by Sergey Shinderuk (@shinderuk), Stephen Frost (@sfrost).