Skip logging script statements for create/alter extension.
PostgreSQL reports the statement text for each statement in the script as the entire script text, which can blow up the logs. The create/alter statement will still be logged.

Since a superuser is responsible for determining which extensions are available, and in most cases installing them, it should not be necessary to log each statement in the script.

Reported by Craig Kerstiens (@craigkerstiens).
Reviewed by Joe Conway (@jconway), John Harvey (@crunchyjohn).