fix a bug that assume the calling order
We cannot assume that the hook function pgaudit_ExecutorCheckPerms_hook is executed after other hook functions. 
Other extensions may be executed in a different order.
This bug can be found from the tag 1.0.0 to the tag 1.5.2 and since the implementation code is changed from tag 1.6.0,we have to make a other patch for it!