Improve logging of compound statements.
This improves the situation where compound statements are sent to PostgreSQL. Previously, all the statements would be output for each audit entry, which is less useful for auditing and wasteful of space. In addition password redaction may redact following statements.

The most significant change for the output is that statement-terminating semicolons are not output. This is probably a net benefit overall but does represent a change from prior output. Leading/trailing whitespace is also stripped.

Suggested by @nachoalonsoportillo, Andres Freund (@anarazel).
Reviewed by Douglas J Hunley (@hunleyd).