Skip auditing of relations without permissions.
Relations without permissions should be skipped since they represent either optimizer artifacts or partitions. Normally they do not cause a problem, but in some cases they lack a distinct mem context and cause the stack to get out of sync. It is also more efficient to skip them and this solution is similar to what was done for PostgreSQL 16.

This eliminates a duplicated log entry that was also fixed in PostgreSQL 16.

Reported by @wzbnb001, Alexander Kukushkin (@CyberDem0n).
Reviewed by Alexander Kukushkin (@CyberDem0n).