Fix search_path for trigger functions.
When quoting the value for the search_path in create function the whole string is treated as a single schema instead of multiple individual schemas resulting in the effective search_path being pg_temp, pg_catalog, "pg_catalog, pg_temp".

Since all functions and tables in the code are schema-qualified this does not result in a behavioral change so does not need to be back-patched. However, it still makes sense to fix it as a safeguard against future changes that are not schema-qualified and because it is the right thing to do.

Reported by Ronan Dunklau (@rdunklau), reviewed by David Steele (@dwsteele).