Free unused command strings.
Unused command strings are currently left in the calling context rather than being freed immediately. This is not a big issue since the context will be freed and the command string along with it but it seems like a good idea to explicitly free these strings in case they are large.

There is no great need for a back patch since the command string is only modified in rarely called commands and it does get freed.