Error when required hooks are skipped.
If the initial hooks (ProcessUtility or ExecutorCheckPerms) that setup an audit event are skipped by another extension then later hooks will be left without an event to update and log, which leads to a segfault.

In this case an audit event will be lost and it seems incorrect to just do that silently. The current behavior of segfaulting is not OK so instead throw an error in these cases.

If the user wants to proceed then they will need to disable auditing, assuming they have permission to do so.

Reviewed by Tristan Partin (@tristan957).