Use strpbrk() instead of strstr() where applicable.
There is a place in pgaudit.c that checks if a string contains characters c1...cN, by calling strstr() for every character in this sequence. This piece of code can be simplified by replacing these calls with just a single call to strpbrk(). In addition, this patch reorders the list of characters in a corresponding comment to make the order consistent with the actual code.

Reviewed by David Steele (@dwsteele).